Rename View to HTMLLocalization
